Move some tests from workflow to software package
